+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - **BREAKING** feat!: remove `Update` and `Complete` typedefs
 - feat: update `FlowController` to have a public constructor
 - feat: `FlowBuilder` requires either `state` or `controller`
-- feat: `FlowController` exposes `addListener`, `removeListener`, and `dispose`
+- feat: `FlowController` exposes `state`, `addListener`, `removeListener`, and `dispose`
+- fix: support hardware back functionality on Android within nested flows
 
 # 0.0.1-dev.9